+++ crates/pdx-rules/src/rulec.rs
@@ -1288,7 +1288,7 @@ mod tests {
         assert_eq!(source_model.file_categories.len(), 124);
         assert_eq!(source_model.symbol_descriptors.len(), 2666);
         assert_eq!(source_model.records.len(), 12_975);
-        assert_eq!(source_model.semantic.rules.len(), 8_311);
+        assert_eq!(source_model.semantic.rules.len(), 8_314);
         assert_eq!(source_model.semantic.enum_values.len(), 72);
         assert_eq!(source_model.semantic.type_root_keys.len(), 7);
         assert_eq!(source_model.semantic.type_root_scopes.len(), 2);